pyproject.toml
[project]
name = "hierfinrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "click==8.3.1",
    "distro==1.9.0",
    "exceptiongroup==1.3.1",
    "h11==0.16.0",
    "hf-xet==1.3.2",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==1.6.0",
    "jiter==0.13.0",
    "markdown-it-py==4.0.0",
    "mdurl==0.1.2",
    "networkx>=3.4.2",
    "numpy>=2.2.6",
    "openai==2.26.0",
    "packaging==26.0",
    "pandas>=2.3.3",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "python-dotenv==1.2.2",
    "pyyaml==6.0.3",
    "regex==2026.2.28",
    "rich==14.3.3",
    "safetensors==0.7.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers==5.2.3",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "tokenizers==0.22.2",
    "torch>=2.10.0",
    "torch-geometric>=2.7.0",
    "tqdm>=4.67.3",
    "transformers==5.3.0",
    "typer==0.24.1",
    "typing-inspection==0.4.2",
]
